# openapi.yaml
...
openapi: 3.0.3
info:
  title: Flowcode Builder Autoren-API
  description: >-
    Lokale Autoren-API des flowcode-website-builder (Next.js Dev-Server).
    Bedient von Menschen (Editor-UI), printed CLIs/MCP-Servern (Printing
    Press) und Agenten (Claude Code, Paperclip-Worker). Schreibende
    Operationen sind explizit; der Import ist dry-run-by-default.
  version: "1.0.0"
  x-auth-type: none
  x-runtime: "next dev only (127.0.0.1:3113) — nicht im Static-Export-Deploy"
servers:
  - url: http://127.0.0.1:3113
    description: Lokaler Dev-Server (Autoren-Maschine)
paths:
...
  /api/abbild:
    post:
      operationId: abbildAktion
      summary: "Bestand: Abbilder des Grafik-Editors (aktion-Dispatch)"
      description: >-
        Historische Sammel-Route des Grafik-Editors (app/api/abbild/route.ts).
        Aktionen: liste | lade | speichere | loesche | standard ("Als
        Standard setzen" schreibt grafik.config.json).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [aktion]
                  properties:
                    aktion: { type: string, enum: [liste] }
                - type: object
                  required: [aktion, name]
                  properties:
                    aktion: { type: string, enum: [lade, loesche] }
                    name: { $ref: "#/components/schemas/SeitenName" }
                - type: object
                  required: [aktion, name, abbild]
                  properties:
                    aktion: { type: string, enum: [speichere] }
                    name: { $ref: "#/components/schemas/SeitenName" }
                    abbild: { $ref: "#/components/schemas/GrafikSetup" }
                - type: object
                  required: [aktion, abbild]
                  properties:
                    aktion: { type: string, enum: [standard] }
                    abbild: { $ref: "#/components/schemas/GrafikSetup" }
      responses:
        "200":
          description: Aktionsabhaengige Antwort (liste/lade/speichere/loesche/standard)
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/Ungueltig" }
        "403": { $ref: "#/components/responses/FremderUrsprung" }
        "404": { $ref: "#/components/responses/NichtGefunden" }
...
components:
  responses:
    Ungueltig:
      description: Ungueltige Eingabe
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Fehler" }
    NichtGefunden:
      description: Nicht gefunden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Fehler" }
    FremderUrsprung:
      description: CSRF-Gate — fremder Browser-Origin abgelehnt
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Fehler" }
  schemas:
    Fehler:
      type: object
      required: [fehler]
      properties:
        fehler: { type: string }
    SeitenName:
      type: string
      maxLength: 100
      description: >-
        Reiner Dateinamen-Baustein (kein Pfad, kein "..", keine
        Windows-Sonderzeichen; Unicode-Buchstaben erlaubt) — Sanitisierung
        in lib/api/server-helfer.ts (saubererName).
...
    GrafikKeyframe:
      type: object
      required: [scrollY, x, y, scale, opacity, rotation]
      properties:
        scrollY: { type: number }
        x: { type: number }
        y: { type: number }
        scale: { type: number }
        opacity: { type: number }
        rotation: { type: number }
        easing:
          type: array
          items: { type: number }
          minItems: 4
          maxItems: 4
        srcOverride: { type: string }
    Grafik:
      type: object
      required: [id, name, src, art, breitePx, z, keyframes]
      properties:
        id: { type: string }
        name: { type: string }
        src: { type: string }
        art: { type: string, enum: [bild, lottie, video] }
        breitePx: { type: number }
        z: { type: number }
        modus: { type: string, enum: [loop, scrub, still] }
        stumm: { type: boolean }
        versteckt: { type: boolean }
        gesperrt: { type: boolean }
        spiegelX: { type: boolean }
        spiegelY: { type: boolean }
        keyframes:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/GrafikKeyframe" }
    GrafikSetup:
      type: object
      required: [grafiken]
      description: Export-Format des Grafik-Editors (components/grafik/grafik-types.ts).
      properties:
        name: { type: string }
        gespeichert: { type: string }
        meta:
          type: object
          properties:
            viewportW: { type: number }
            viewportH: { type: number }
            docH: { type: number }
        grafiken:
          type: array
          items: { $ref: "#/components/schemas/Grafik" }
        pool:
          type: array
          items: { type: object }
        uebernommen:
          type: array
          items: { type: string }